Update some fields type. Fix tests

+++ pa-ca/src/main/java/com/paca/paca/business/model/Tier.java
@@ -2,6 +2,9 @@ package com.paca.paca.business.model;
 
 import lombok.*;
 import jakarta.persistence.*;
+
+import java.math.BigDecimal;
+
 import com.paca.paca.statics.BusinessTier;
 
 @Entity
@@ -16,12 +19,15 @@ public class Tier {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
     private Long id;
+
     @Enumerated(EnumType.STRING)
     @Column(name = "name")
     private BusinessTier name;
+
     @Column(name = "reservation_limit")
     private int reservationLimit;
+
     @Column(name = "tier_cost")
-    private float tierCost;
+    private BigDecimal tierCost;
 
 }
